requirement.txt
asgiref==3.11.0
Django==5.2.8
django-grappelli==4.0.3
django-jet-reboot==1.3.10
pillow==12.0.0
sqlparse==0.5.3
